Update configuration for GOV 51.
Note that the configuration has been copied over from the gov50 branch
since it should be more or less the same. Just renamed the course.

Since the GOV 51 image has not been deployed yet, the form.yml
references the GOV 50 image. This will be updated when the image has
been deployed.